PEPPER-849 OSTEO1 user creation API for test support
